21:16

+++ Ingrediente.java
@@ -10,7 +10,7 @@ public class Ingrediente {
 		precio = 20.50;
 	}
 	
-	public Ingrediente(String nombre, int cantidad, float precio){
+	public Ingrediente(String nombre, int cantidad, double precio){
 		this.nombre = nombre;
 		this.cantidad = cantidad;
 		this.precio = precio;
